requirements.txt
# Core Flask dependencies
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
Flask-Mail==0.9.1
Flask-WTF==1.2.1
WTForms==3.1.1

# Database
SQLAlchemy>=2.0.36,<2.1
psycopg2-binary==2.9.9; sys_platform != 'win32'
alembic==1.13.0

# Security & Authentication
python-dotenv==1.0.0
bcrypt==4.1.1
PyJWT==2.8.0
email-validator==2.1.0

# Serialization & Validation
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0

# Background tasks
celery==5.3.4
redis==5.0.1

# Testing
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0; python_version < '3.13'
faker==20.1.0

# Development
python-dateutil==2.8.2
pytz==2023.3

# API Documentation
flasgger==0.9.7.1

# File handling
Pillow==10.1.0; sys_platform != 'win32'
reportlab==4.0.7

# Analytics & Charts (skip on Windows)
pandas==2.1.3; sys_platform != 'win32'
numpy==1.26.2; sys_platform != 'win32'

# Production server (skip on Windows)
gunicorn==21.2.0; sys_platform != 'win32'
Flask-Caching==2.1.0
Flask-Limiter==3.7.0

# Linting & Formatting
flake8==7.1.1
black==24.8.0
